+++ CHANGELOG.md
@@ -20,6 +20,9 @@ All versions prior to 0.9.0 are untracked.
   [#1402](https://github.com/sigstore/sigstore-python/pull/1402)
 
 
+* Added a `RekorV2Client` for posting new entries to a Rekor V2 instance.
+  [#1400](https://github.com/sigstore/sigstore-python/pull/1422)
+
 ### Fixed
 
 * Avoid instantiation issues with `TransparencyLogEntry` when `InclusionPromise` is not present.